+++ web-frontend/modules/database/components/table/TableJSONImporter.vue
@@ -190,9 +190,15 @@ export default {
 
       data.unshift(header)
 
-      this.values.data = JSON.stringify(data)
-      this.error = ''
-      this.preview = this.getPreview(data, true)
+      try {
+        this.preview = this.getPreview(data, true)
+        this.values.data = JSON.stringify(data)
+        this.error = ''
+      } catch (error) {
+        this.values.data = ''
+        this.error = error.message
+        this.preview = {}
+      }
     },
   },
 }