+++ web-frontend/modules/database/mixins/importer.js
@@ -34,10 +34,37 @@ export default {
         remaining = data.length - rows.length
       }
 
+      this.validateHeader(head)
+
       head = fill(head, columns)
       rows.map((row) => fill(row, columns))
 
       return { columns, head, rows, remaining }
+    },
+    /**
+     * Validates that the uploaded field names are unique, non blank and don't use any
+     * reserved Baserow field names.
+     * @param {*[]} head An array of field names to be checked.
+     */
+    validateHeader(head) {
+      const headSet = new Set()
+      // Please keep in sync with src/baserow/contrib/database/fields/handler.py:30
+      const RESERVED_BASEROW_FIELD_NAMES = ['id', 'order']
+      for (const column of head) {
+        const trimmedColumn = column.trim()
+        if (trimmedColumn === '') {
+          throw new Error('Blank field names are not allowed.')
+        }
+        if (RESERVED_BASEROW_FIELD_NAMES.includes(trimmedColumn)) {
+          throw new Error(
+            `${column} is a reserved baserow field name and cannot be used.`
+          )
+        }
+        headSet.add(trimmedColumn)
+      }
+      if (headSet.size !== head.length) {
+        throw new Error('Field names must be unique.')
+      }
     },
   },
 }